Code to convert Geolab iob file to a DynaML.
Code to convert Geolab iob file to a DynaML.
iob file is specified in the arguments line.
The code also has options to break the iob network into discrete networks and reweight these for propogation of uncertainty. However for this function to be available. The PLO / PLO record must contain extra infrormation that supplies the a-axis, b-axis and Az of the error ellipse on fixed stations.